Add a minim bit size for subword atomics
As upstream is currently working on a complete implementation of
atomics, we only need a stop-gap solution that allows us to use 32/64
bit atomics when possible and defer everything else to library calls.

Due to the way the mechanism works, this means that atomic loads and
stores of subword size will also use library calls. Truth is that these
can be implemented with instructions already but it seems worthless,
given the transient nature of this change, to tell load/store apart from
RMW instructions. So the tests atomics.ll and atomic-load-store.ll have
to change to reflect the worse code generation.