Add -latomic to the link step
As a stop-gap solution, when the compiler uses atomic runtime calls,
make sure we link with -latomic

We will be able to remove this once the compiler does load subword
atomic load/store. When this happens a number of tests in LLVM will
start failing so we will notice.

This fixes EPI/llvm#6