Reserve the FP register when using EPI
Unfortunately, due to the design of LLVM, we realise too late that there
are spills/reloads of vector registers. The register allocator queries
the reserved registers very early on its operation so it ends using FP
even if we need it when we're spilling EPI registers.

So for now, reserve it unconditionally. It seems doable to write a
post-RA pass that undoes this.