Allow a more relaxed alignment for load/store of vector types
Both clang and the V-extension only require an alignment as strict as
the alignment of the element type. But currently LLVM assumes a stricter
SIMD-like alignment is required (= sizeof(element-type) * num-elements).

This causes the load/store, albeit legal, be expanded and then
unexpected paths of the legalizer, not ready yet for scalable vectors,
kick in.

This is mostly a band-aid for -O0 than a real fix.