Use the right tblgen
We were using llvm-tblgen which given that this is clang may be a
questionable choice. But worse than that we were using it in a way
that it as being found in the PATH.

Use a cmake generator expression with the right target (clang-tblgen)
so it uses the right clang-tblgen, no matter it is an external one or a
built-in one.

This commit fixes ci-infrastructure#13