Add a mechanism to document builtins
Although these builtins are of internal nature and at some point
will be replaced by a proper EPI/V-extension header, we need a way
to document them.

Remove the existing documentation class from epi_builtins.td and
include that file from epi_builtins_doc.td to link the documentation
to the definition of the builtin.

Then document the builtins and emit something for them. Currently a form
of Markdown that looks reasonable in gitlab. Other documentation
backends (like AsciiDoc) may be of interest in the future.