Overhaul masks and implement mask binary operations
So far masked instructions had an explicit dependency with the physical
register V0. It is better to adjust that so we create a subclass of
EPIVR, EPIMaskVR, that includes just V0. This way we can naturally
represent the input dependence with a mask register (currently just V0)
as a normal input.

Also RISCVInstrInfoEPI was representing as a parse operand the actual
encoding for masks (a 0 bit, actually). It is better to give this
operand a value of the EPIMaskVR register class. This means that
instructions that force a mask operand (like vmerge and vadc) can just
use an operand of kind EPIMaskVR (so they will parse 'v0' automatically).

This means that now "vectormask:$vm" is a normal register operand
instead of just a random bit in the instruction. Note that this applies
only to "real" instructions. Pseudo instructions already had vector
masks as input operands. The novelty now lies that
RISCVExpandPseudoInsts will correctly forward the mask operand to the
real instruction.

For instructions where the mask is '<mask-operand>.t', this change
updates the parsing, validation and decoding routines so they compute a
RISC::V0 when "v0.t" is parsed.

This change means that now our register classes are a bit more
complicated so many direct comparisons won't work and will now require
hasSuperClassEq.

Also instructions vmerge, vfmerge, vadc and vsbc now they have an
explicit mask operand. They didn't have it before because V0 was
included in their Uses.

Finally this change adds a couple of patterns for zext and trunc masks
For EPIVR, zext is implemented using vmand.mm as suggested in the V-ext
specification. trunc is also implemented using vmand.mm, however it
should be possible to simplify this case when the result of the trunc is
used as the mask of a masked instruction (but only for EPIVR). It should
be possible to add an extra pattern in this case. However one wouldn't expect
this sequence be generated explicitly unless the mask is reinterpreted from
an integer vector. For EPIVR{2,4,8} trunc and zext are implemented using
both arithmetic and relational operations.

Finally this patch also reenables patterns for the mask binary
operations.