Fix zero extension in optnone
DAGCombiner may want in some cases, to build an "and" with a constant
vector of all ones.

ISD::BUILD_VECTOR can be generated by the DAGCombiner to represent
splat vectors so let's lower it to RISCVISD::VBROADCAST.

This patch also adds a pattern for a whole-register binary and so the
code can be generated.