Clang/LLVM: Fix CodeGen of references.
The new way behaves as follows:
- References are captured by saving the load of them in the bundle. That is, the referenced
value.
- Global constant references were simplified by CodeGen to the referenced value. This is
skiped since we want to keep the dsa of the reference and the value referenced different.
If we allow this, we'll end up having two data sharings for the same Value.
E.g.
   (global)
   int i;
   int &ri = i;
   task private(ri) shared(i)
   {}
---
QUAL.OSS.PRIVATE(@i), QUAL.OSS.SHARED(@i)
The first is the simplification made by CodeGen

In order to match the referenced Value emited for DSAs we have a map to
reuse the Address already emited, making all other clauses emission
and task body emission consistent.