Expand EPI pseudo-instructions in AsmPrinter
This is required when we expand a LMUL > 1 instruction into multiple
LMUL = 1 instructions.

For instance, if this expansion is done at the RISCVExpandPseudoInsts
phase, a pseudo using a V2_2 register could be expanded to multiple
instructions using V2 and V3, but the MachineInstr verifier will
complain about the use of V3 (since other definitions to V2_2 will have
been lowered to V2, and there will be no definition of V3).