Rework a bit Shaping expr parsing
Instead of trying to parse a parameter declaration list, try to parse an assignment expression.

The old way doesn't work because we can have '[x](int *)p' and when (int *) is a valid
parameter declaration list, but in reality it's a cast.

The new way has also some caveats, because allows to write '[x](p = 0)' for example