Refactor dependencies (See full message for details)
The new dependency/reduction bundle has the form (key, compute_dep, args...) where:
1. key is the base of the dependency, a DSA.
2. compute_dep is the function that will perform the dependency info needed by nanos6
3. args are the parameters to be passed to 2.

To gather the parameters to be passed to CodeGenFunction of compute_dep we add a new visitor
before the OSSDependVisitor. This visitor visits in the same way as the other to detect
the number of dimensions and the dependency type. Additionaly, it visits all the subexpressions
to find the parameters to be passed.

We need all the VLASize expr. because CodeGen seems to need them. Also there is a case where
the user may write 'in(vla[sizeof(vla)]', that also needs all the VLASizes.

For Shapes, we gather the variables involved in the shape. OSSDependVisitor builds the
whole expressions.

In LLVM we now can check that all the info in the dependency bundle is a DSA or a capture and
perform an assert.

The address translation is only performed for reductions

Closes #62

* Additional fixes

- Global variables are not used as compute_dep func. parameter since CodeGen emits them directly.
  To solve this, we make the map of references a map of captures, that will have other stuff apart
  from references. For now we add globals here too.
- There was a bug when using parentheses in a dependency [i.e. in( (a) )]

* Updated tests

- Added a test for sizeof vla and vla shapes with subscripts
- Remove task_layout_with_unpack.ll and task_loop_iter.ll
  because we do not have unpack instructions anymore

* Addicional Note 1

- Skip non_odr_use expressions in OSSClauseDSAChecker.
- Skip annotation of non_odr_use expressions for compute_dep function creation
    - Only skip non_odr_use_unevaluated, we want to annotate
      dsa for non_odr_use_constant (i.e. global reference)

* Addicional Note 2

Workaround to make pointers/basic types/arrays/structs to work with reductions

The main problem here is that we do not have any information about the C/C++ type
we are processing. Because of this it's difficult to establish a criteria to decide
if we have to pass the value of the DSA (pointers) or the storage (others).

The easiest way to make this to work is to check the returned type by compute_dep. If
it's different from the one we passed we assume a load has been performed, so we're
processing a pointer.

* Addicional Note 3

Mark success_task_25 as XFAIL since it's not supported.
This worked because the global value was somehow used, but it is not correct