Fix #73. Store the translation result into task args. See more.
At this moment we have decided that address translation is going
to be performed only for reductions and device dependencies.

This is because we only support one indirection translation, so

    int **p;
    in([10](p[3]))

will not work.

Maybe in the future we can add translation for those exprs. that
can be translatable, although in smp is unnecessary...

Fix reduction init function.
Closes #73