Fix bad diagnostic when having a constexpr template function call in OmpSs-2 context
The old method to detect a return instr. inside task body does not work because
Scope is used only in parsing, not in instantiation.

To make it work properly we push a fake FunctionScopeInfo with a flag to detect
we are in task body context.
Also, since a lambda tries to propagate captures to outer ones
we have to skip our fake FunctionScopeInfo.

Closes #96 #97