Go back to the old way to handle this in compute_dep. See more.
The refactor was putting 'this' as a arg/param even when it's not used
to compute the dependency.

This is a problem in task outline where we firstprivatize 'this', so
we end up with an argument in compute_dep that is not in a data-sharing.

In this commit we only make 'this' to appear in compute_dep when it's used.