[OmpSs-2] Refactor of analysis and transform passes. See more.
Since the current codebase is a bit hard to extend to support new directives
we have changes some code to make it easier in the future:

The important changes are:
- We only care about if the directive is a marker or a region entry/exit.
- Regular dependencies, reductions, and multideps are now the same structure/hierarchy.
- Introduce a DirectiveEnvironment that gathers all the ompss-2 bundle information and
  performs some checks.

With this, we only have to pass a post order directive array to transform an lower them
in the specified order.

Also, since deps/multideps/reductions share the same structure we can reuse a lot of code,
and avoid to handle manually the dependency kind (i.e. in, out, inout...)

We have refined directive auxiliar function construction too
(i.e. dependencies/constraints...) and we only build them when it is necessary.