Allow promotion to shared only in subexpressions.
Before, any attempt to make shared a variable was allowed,
but that is not correct.

For example:

    int a;
    int *p;
    \#pragma oss task reduction(+: [a]p) shared(a, p)
    {}

'p' cannot be shared since it will replace the original value
by the reduction storage.
'a' is allowed to be shared.

WARNING: this relaxation can make the program to segfault, since
the shared variable may not exist when evaluating the reduction/dependency

For example:

    void foo(int *p, int a) {
        \#pragma oss task reduction(+: [a]p) shared(a)
        {}
    }

When exiting the function 'a' does not exists anymore.

NOTE: Took the oportunity to refactor the DSA clause visitor.