Rework mechanism to detect conflicts between dependencies and reductions.
We introduce the concept of strong and weak restriction:

A strong restriction is a reduction over a variable that is the symbol.
A weak restriction is a variable used in reduction/dependency subexpressions
or a dependency symbol.

So, reduction(+: x) is a strong restriction in 'x', but is a weak restriction
in in(x, v[x], [x]p, ...) or reduction(+: [x]p).

We emit a diagnostic for strong restrictions used in other weak restriction places.

Closes #85