Build allocas in a one time executable place. See more.
This is an issue being from the beginning of the lowering implementation.
Alloca instructions have to be "executed" once, so having one in a loop
makes the stack to increase every iteration ending up in a stack smashing.

This issue happens in two places:

- When creating the ".remap" variable used by discrete multideps:
FIX: put the allocas in the entry block of the function we are.

- When creating nanos6_create_task needed allocas.
FIX: Since we may have task nesting, the solution is to annotate
the allocas an move them to the function entry block that has its uses.