Emit a warning on task outline function redeclaration. See more.
Clang uses the last declaration to build the call node.

Since we allow to specify that both function declaration
and definition are a task outline we may find some program whose
function declaration is not a task outline but the definition is, and
the other way around.

Also it's allowed to specify that both declaration and definition are task
outline, but the clauses may be different.

This commit fixes #84