Fix race when the team finishes
If there are unshackled threads executing tasks of a team, the team
cannot finish before all those unshackled threads have finished too.

To do this, a team is open for unshackled threads and it stops being
open for that once the team is finishing. Then we proceed to wait for
any unshackled to finish. The idea is that no other unshackled attempts
to pick work from the team, which will be invalid after it finishes.