Refactor of DSA deduction pipeline. See more.
Due to our necesity of promote DSA from firstprivate to shared
when a variable is used in cost/priority/onready or task body I've
refactored the whole mechanism a bit to support this.

The new pipeline works like this:

Let be IsBase a bool that indicates that a variable is not used in a subexpression.
Example:
  i = 4 // 'i' is base
  array[i] // i is NOT base

Let be Ignored a bool that indicated that a variable must not be analyzed
Let be a Promotion a change of DSA from firstprivate to shared only when

1. Explicit DSAs have always IsBase=true and Ignored=false
2. Dependencies assign an implicit DSA if not defined yet or
   a promotion
3. Cost/Priority/Onready and task body NOW can do a promotion too.

* IndVar of a loop directive is always Private and IsBase=true and Ignored=true.
  This is because we want to allow the user to use it as a dependency
  and don't want to emit DSA conflicts
Example:
  pragma oss taskloop in(array[i])
  for (int i ...) {}

* Multideps are always Private and IsBase=true and Ignored=true
  This is because we don't want to emit DSA conflicts when used in a multidep expr.
Example:
  pragma oss task in( { v[i], i = 0;10 } )
  {}

Addicionaly, the final implicit DSAs info needed to build the clauses has been moved to the stack,
instead of having that info in every visitor.

Closes #132