Fix unshackled victim's choice code.
The victim's choice code is in a bad place for an unshackled thread because
we can only fall there when victim_tid == -1 and new_victim = 0. This is
only possible when:

1. We have no tasks in our deque
2. We are in 1., choose a victim to steal, execute the stolen task and
   new tasks are created, so use_own_tasks=1 and new_victim=0.
   And now we execute all the tasks of our queue, so victim_tid=-1.

Unshackled thread gets stuck in execute_tasks when steals all possible
tasks, but any of them creates more tasks,
so new_victim=1 and victim_tid != -1.

Choosing the new victim when checking the exit condition guarantees
progress.