[1/2] Fix barrier-like behaviour when many task_teams end at the same time.
We have seen in paraver that many parallels ending at the same time wait for
an unshackled to release its list of task_teams lock. This is because
each parallel has to remove its task_team from every unshackled.

To minimize the time with the lock taken we add a wait for unfinished
unshackled threads in the master task_team, this way an unshackled only has
to take the lock until the counter is incremented.