Make task team of a serial team reusable and fix master thread wait(unshackleds)
tt_active is set as FALSE once a __kmp_task_team_wait is performed. In serialized parallels
it is not set to TRUE again so unshackleds cannot execute tasks there. We fix by setting tt_active=TRUE
every time we performa a serialized parallel.

Also, it is not good to do a flag wait after putting the task team of the master to NULL because
the thread may go to sleep and never resume. We fix this by doing the wait before the desactivating
the task team