Rework multidep loop building. See more.
Until 4319b1b95cba we had to perform the iterator computations
at every step of the loop, because the user could use the same iterator
in its loop expressions.

Example: {v[i], i=f(i);g(i) }

This is not allowed anymore so we can just compute the stuff at the beginning
of the loop. Remap has to be recomputed each time because it's used
in discrete multideps to index the discrete deps array.