Add initial support for taskiter (for|while). See more.
Here is a description of the implementation.

Clang:

Taskiter (for) is modeled as a loop directive, like
taskloop.

Taskiter (while) is more permissive, allowing any C/C++
condition in the while loop. This includes lambdas and
condition using pointers.

The loop iterator is firstprivate instead of private.
Because in the llvm transform is needed.

- Unroll clause is an integer > 0

NOTE: At this time the loop iterator is not usable in
taskiter (for) clauses.

NOTE: At this time while condition used variables
are not specially handled to make them shared variables

NOTE: There is no support for stride() clause yet.
Its implementation is the same as unroll clause

LLVM:

Taskiter is lowered in the unpack function as a simple
body without the loop. In particular, taskiter (for)
includes the body and emits the code to increment the iterator.
In taskiter (while) only the body is emitted.

The reason is because the runtime will have the amount of
iterations to execute (for version), or a task_info field
with the function that evaluates the condition.

NOTE: In contrast to Clang, Taskiter is NOT a loop directive
because the lowering in the unpack function differs.

Special rebase notes:

  Remove apparently unused code to clear OSSLateParsedToks

  In 35ff50076f43 removing this code results in the test error.
  I add the test to keep track we do not introduce the bug again.

Reference #159