Add debug support in OmpSs-2 directives. See more.
This implementation reemits debug info in the task body and
hacks CodeExtractor to not remove intrinsics using task params.

A few notes to justify this implementation:

VLAs types cannot be reused since they depend on
vla size variables. Although vla sizes are captured in OmpSs-2
directives there is no easy way to build the vla type containing
the vla sizes correctly since they are 'i64' values.

I investigated the possibility of using llvm.dbg.value intrinsic
to make it work, but and additional instructions using the value
are required (doing an alloca and a store there for example).
If the value is ONLY used in the intrinsic CodeGen does
not emit anything into the binary.

Then, for simplicity (and since clang already does it), VLAs
will be seen as unbounded arrays.

'this' and references are supported by emitting the non referenced
type debug info instead of the regular one. This is because we capture
the load of the allocas, which are the locations of the referenced
variables.

Fixes #59 #143