Extend label clause to support. See more
Our label implementation was limited to a string
per task_info. Clang allowed to specify non constant
strings but it was a hacky feature. Also don't work
because every task creation will overwrite the task_info label.

This new implementation keeps the constant string for the task_info
(task type), and introduces a new string that is passed to the
runtime in nanos6_create_* functions. This new string will be
instantiation dependent and will be used to improve tracing.