Fix a TAMPI bug. See more.
A TAMPI task is a regular task that is converted
to detached. This causes a incongruence between
the task allocation and the task free using proxy
tasks mechanism.

In a serial team, the parent task td_incomplete_child_tasks
counter is only incremented if the allocated task is detached.
In our case, the task is resolved to detached later, so we want
to increase this counter too.

The solution is to increment the counter at allocation time
checking the tt_found_proxy_tasks flag that TAMPI has already set
in the MPI_Init(). In regular detached tasks it is redundant,
because the same flag will be set to one before this code,
becuase the task is DETACHED.