Improve OmpSs-2 Stmt allocation code to avoid errors. See more.
This commit abstracts memory needed for the OmpSs-2 stmts info.
This code has been borrowed from OpenMP

commit 0af7835eae698330f6362220ff1fcbfe92516de8
[OPENMP]Redesign of OMPExecutableDirective/OMPDeclarativeDirective representation.

We keep NumChildren parameter because it could be useful for newer directives
that need some additional info. In theory we should use this for loop directives
to store all the iterators, loop bounds... but we have other values appart from
'Expr *', like 'bool'. Storing different sizeofs is not really convenient so
I kept the original mechanism for this case.