Import OpenMP 5.2 atomic directive and clauses (except for 'compare' clause). See more.
Fow now, and since OpenMP has already done the work for us let's take profit of it.
'compare' clause is not as easily to copy since it's implemented in OpenMPIRBuilder.
We're a bit outdated in this part, so I'd prefer to update our OmpSsIRBuilder before
adding support for it.

Also, we have not decided if it'd be better to implement the
directive in LLVM Transform phase or not, and how to do it
portable enough.

Closes #172