FIXME: Add workaround to handle array descriptors (fortran POINTER) in address translation. See more.
Until now we were safe assuming a firstprivate dependency could only be a pointer (C/C++). Now
with fortran POINTER array descriptors this is not right. The workaround is to check if the address
to be translated is a struct and load the first field address if so.

A more generic solution language dependent (a function provided from the Frontend for example)
would be better, but for now I think we're fine with this.

The issue tracking this is #176