Capture vla expressions and generate compute.dep based on them. See more.
We build a fake symbols in the OSSDependInfoGathering as replacement of original symbol
when emiting vlas.

We have to map the original value to the compute.dep function parameter
because generating a expression will reach the original symbol.

Similarly, for our dependency analysis we have to use the modified symbol

NOTE: Creating an auxiliar function requires all cloned symbols to be dummy.
      The reason is because if they are not dummy some initialization will be
      triggered like allocatables or pointers.