Add support for ASTImport to report errors.
This happens in some unsuported code, unrelated to my TFG. Clang
currently is not quite able to import all C++ code, and fails with some
code from std::cout (for example). This will be fixed when C++20 modules
are fully supported. In the meantime, error more gracefully (and tell
the user where's the problem).