Fix for device pure task call order. See more.
We found that a device pure task that includes references to variables
that are not params of the function alters the task_args order.

By adding a new bundle QUAL.OSS.DEVICE.CALL.ORDER we put these
values just after the ndrange values. This way the runtime
is safe to pass them directly when executing the kernel.

I've added checks in the llvm tests to verify the global variables
data to be more strict.