Add -fopenmp-nosv. See more:
For my state of mind, I prefer to keep the openmp
llvm ir tests passing. This means we have to keep
the original CodeGen so I add this new flag to enable
the CodeGen for our nosv runtime.

This makes clang available to generate right code for any
other openmp runtime. Our openmp runtime only will be able
to work with -fopenmp-nosv. Newer commits will add
the changes in the runtime side