Fix __glibcxx_assert_fail when building debug release
This function is defined in c++config.h in gcc (it is part of
libstdc++). On gcc 11.3.0, this symbol is defined in the mentioned
header file if the GLIBCXX_ASSERTIONS define is set at compile time.
However, on gcc 12.3.0, this symbol is declared as extern if the define
is set. The compiler fails to link the object file or library where it
is defined, and the symbol on kmp_csupport.o remains undefined.

This https://gcc.gnu.org/git/gitweb.cgi?p=gcc.git;h=a54ce8865a885bca5ab9c4aa6ec725cd13c09901 introduces the change.
This issue has been mentioned before in flang https://reviews.llvm.org/D142279 and decided to undefine GLIBCXX_ASSERTIONS to avoid an "unwanted dependency with stdc++".

This patch follows the flang approach and disabled stc++ asserts.