Fix nosv_self inconsistency with th_nosv_task. See more.
We were trying to track the nosv_task being executed. This
is needed because workers are wake up by other threads
in the team and they need the nosv_task where the victim
is paused.

In openmp only programs works because there are only openmp
tasks. If other entity creates other nosv_tasks (NODES)
this does not work since the current nosv_task is changed
without tracking.

The solution appears to be simply get the current nosv_task
just before pausing the thread.