Some things regarding rpath. See more:
We want to add rpath to openmp too so we do not rely
on LD_LIBRARY_PATH. This makes things easier for tools (and nix)
to just use the compiler to run tests, instead of trying to guess
where is openmp library.

If we add rpath to openmp we will have a problem with the slurm modules.
This is because both nosv and omp are in the same place. We added
NOSV_HOME rpath after OpenMP rpath, and this means the path where
OpenMP is has priority. We do not want that.

This patch adds rpath to OpenMP and puts nosv rpath before it.