Fix passive. See more:
I've found that some openmp tests do not finish when setting PASSIVE
as a default. This happens when a thread that is not the master
creates task. There is a window between the last task execution
and the worker thread that created that task where the worker do not
decrement the unfinish_threads counter and goes to sleep.

This makes the master thread wait until unfinish_threads to be zero
forever.

In ACTIVE or in a naive openmp this never happens. In the first case
because the thread does not go to sleep and in the second case
because the master thread will wake up the worker eventually
by the "preventive deadlock policy"

Closes #209