Fix race condition of enabling tasking by initializing it just after allocation. See more
This patch changes the OMPT behavior as described in #210 but we're fine with it
because any regular openmp runtime will do the same in a tasking program.

kmp_num_teams.c has been marked as UNSUPPORTED because there is an issue
related with reusing the parallel team in a teams directive. This is not
related with us.

Fixes #210