Add a mechanims to leave always a workers that is not Sleepable spinning. See more:
Until now the amount of workers is manages by the number of running workers
and the amount of tasks. This is not enought because we could have more than
one creator and the current mechanism that keeps two additional workers
running may not work. An example is a parallel region where > 2 workers
create tasks.

This situation delays the task execution until a worker reach a taskwait
or a barrier.

This patch forces the first worker that reaches the barrier to be the
manager and it will not go to sleep.

In the previous described example. The creators will be allocating
and submiting tasks and the first worker that reaches the barrier
will be the manager and will wakeup other workers if needed