Move wakeup mechanism after tasks deques access.
There was the possibility that a worker sees that there
is no need to wakeup a new worker but after a new ready
task is submitted and picked by that worker.

Is the task is long enough it could hinder the performance.
This way if a worker was able to pick a task it is safe
the ready tasks counter is >= num_ready_tasks