Guarantee hash table data read/write memory ordering. See more
There is a chance where modifying 'key' and 'hash' that the value of
'hash' is visible to other thread running the find(), but 'key' is not
updated yet. Use acquire/release to guarantee that. We don't care
about failing the find without lock as we will get the lock to do
the find() again.

Took the oportunity to do some refactor