Fix shared extern struct declaration.
The verifier blames about our bundle notation
to keep information about types for our transform
pass. In this case extern struct types are
in LLVM "type opaque".

As this kind of variables are only allowed to be shared
it is safe to assume the LLVM type "ptr" instead the original
type opaque. We don't use type information in shareds

Closes #255